First cut at removing the opt-out language
This was as little as I could manage with respect to turning "opt-out"
to something more general: "preferences" or "expressions of preference".

There are a bunch of other challenges to address that we should probably
look into.  Some of those change the language a lot more, but I thought
that it was a good idea to get rid of the opt-out bias as a first order
of business.