The abstract model
There are a few things here that I'm struggling with, but I'm starting
with the basics.

The first is to better define the abstract model.  That is, each
category of use has a preference that is either "allow", "disallow", or
"unstated".

The next is to give us more space for category definitions.  Right now,
these are pretty short, but we'll need to say more.  I've not modified
text much, but I've added notes about AI and GenAI being subsets of the
preceding categories.

I've moved a few things around as well, to get things into a better
flow.  In preparation for adding more text on exemplary serialization,
search, inference, and other such things.