Define an exemplary format
As discussed, we want to define a recommended format.  This executes on
that.  It uses Structured Fields, which will ensure that we have a very
easy translation to HTTP fields.  It also fits neatly into the other
formats we are considering.

There are some interesting implications that are worth highlighting.
The text I've added goes into some of that in more detail, but just to
list them for visibility:

* This is a sequence of bytes, not a string.  That has implications for
  how it is included in some formats, but as RFC 9651 limits its syntax
  to a pretty small subset of the ASCII range, this is mostly just a
  theoretical consideration, not a practical one.

* Parsing errors in RFC 9651 means that you lose everything.  No
  preference survives.  That means that authoring errors have
  consequences.  This is an artifact of the format being more of a
  computer-to-computer language, rather than being tolerant of human
  error.  I'm not thrilled with this, but recognize the trade-off
  involved.

* Repeated keys mean that only the last one is taken.  Again, not
  thrilled with this, but think that the trade is tolerable.

* Labels are lowercase only.  A lot of the list traffic has started to
  use AI=whatever; RFC 9651 doesn't let us shout.  I think that this is
  fine, but it's another minor ding.

Closes #8.