Switch to using aasvg for the diagram
I've rearranged the diagram slightly to ensure that the categories are
shown in the order we present them later.  I've also removed the "other"
part, which didn't work that well otherwise.

This will get interesting when more items are added, but I have ideas on
that front.

Sorry Paul, I destroyed your lovely hand-crafted SVG here.  I hope
that's not an unacceptable loss.