Removed reference to origin of TDM definition
This pull request proposes to leave the definition of TDM unchanged, but removes the explicit reference that this is based on EU law. This is part based on a review of existing legal definitions of TDM in other jurisdictions that have TDM exceptions in ether copyright law: 

The UK does not have a proper definition in the law itself, but the UK government provides on as part of its Guidance on Exceptions to copyright which is very similar to the EU definition: 

"Text and data mining is the use of automated analytical techniques to analyse text and data for patterns, trends and other useful information."

The other two jurisdictions with TDM exceptions that I am aware of (Singapore and Japan) do not use the term TDM. Instead they anchor the exceptions on “computational data analysis” (SG) or “data analysis (JP). 

Given this it seems sensible to keep the definition of TDM as unchanged.