Added inference category
This is a first attempt at a definition of the inference category. This is taken from the the C2PA's do not train assertion https://opensource.contentauthenticity.org/docs/manifest/assertions-actions#do-not-train-assertion and is based on their definition of `c2pa.ai_inference`. 

I would consider this to be a staring point for discussion and I do not have a strong attachment to the proposed language.