Add a description of how the model is interrogated
This includes the basic "I want to know what the preference for X is"
as well as combination.

This takes a very specific viewpoint with respect to combination.

It says several things:

1. Each statement applies its own rules, completely.
2. Inherited values therefore apply before combination.
3. The least-permissive statement applies when there is a conflict.
4. However, if one statement is defined to override another (as we might
   decide to do with the HTTP header attachment overriding robots.txt),
   then the overridden statement is ignored.

To illustrate with an example, if you have two sources of preferences,
you might have `ai=n` and `genai=y`.

On its own, the first would say that `genai` defaults to `n`.
The second has no effect if they are combined, because `y` is weaker
than `n`.

Similarly, `ai=y` plus `genai=n` also results in `n`, but that
answer is taken from the second.

For #14.