Revise intent statement
This was the hardest part of #31 and I think we could make this text
clearer.  It is a bit too dense, even with what I'm proposing, but I
hope that it's an improvement.

This is a version of the text provided by Timid Robot in
https://github.com/ietf-wg-aipref/drafts/pull/31#discussion_r2100758813
with some additional edits, particularly the last sentence,
where I think we need to be clear that:

1. This shouldn't prejudice decisions about other rights outside of the
   meaning of the preference itself.
2. The no prejudice piece applies even if the preferencing has legal
   effect, outside of the direct and obvious legal effect itself.
3. This isn't necessarily exclusively about copyright.  People have
   privacy rights over content that still apply, even if the preferences
   admit the possibility of use.

I think that this maintains the intent of the original, but I will admit
that the original is dense enough that I might have lost some of the
intent.