Try to shorten the abstract
This is always going to be difficult.  I might have missed the mark
again here, but I really want to see if I can reduce the size of the
abstract and avoid it sounding too much like legalese.  That's a
difficult task because we really need that sort of precision, even here,
were there is a lot more freedom in terms of omitting finer detail.

In the process, I want something that fixes #19.  Looking at the
introduction, I think that we've largely achieved that for the main body
of the text.