A very short applicability section
I'm putting this further down in the document as I want to make sure
that we get the basics out of the way before getting into the heavy
stuff.

Looking for thoughts about whether this fits with the general thinking
that we have so far.  Especially looking for thoughts on whether this is
pitched at the right level as well.  Is there too much detail or too
little?

For #15, but not included to close it.