IANA registry
This enacts the suggested approach for vocabulary extensions I described
in #38.

Still, this was particularly tricky to get the wording even remotely in
the space I wanted and I'm not sure whether I got there.  Input on this
would be appreciated.

Basically, I'm conflicted on this.  Using Specification Required gives
some flexibility for the process to recognize emerging labels without a
lot of procedural overhead, but it seems like this is angling toward
IETF Review as the operative policy.  Should we face facts and just
go straight for IETF Review?  That would allow the advice to be a little
more streamlined, because we'd be able to rely on the standards process
to iron out the kinks.

There's also the option to pair IETF Review with IESG Approval, which
means "get an RFC" or "convince an IETF AD".  That almost amounts to the
same thing in practice (if you can convince an AD of something, one of
those things is publishing an RFC.  Though the RFC still requires IETF
consensus, so it's a somewhat higher bar, but only "somewhat" given how
careful most ADs are).

In any case, this seemed like a reasonable starting point for a
discussion, which is always easier when there is something concrete to
discuss.