Update acknowledgments
The IETF policy is that people participate as individuals, which means
that we do not generally list affiliation when it comes to
acknowledgments.  There's a carve-out for authorship and a few other
things, but this is how we generally operate.

The formatting ensures that people can use their actual name.  Not a
problem for these all-ASCII names, but it might end up being a problem
later.

I've also sorted the list (by first name, which is my preferred approach
once we have to deal with people who put family names first and all
that).

There are probably a few other people that we might add here, though I
want to be clear about the level of contribution before doing that.
Both Brad and Felix have been very helpful on-list lately.