A reframing of search
As discussed on the list recently.

This does not include Brad's suggestion of "for the exclusive purpose
of".  That is a good change, but I think we need to address that
separately.  That is, we need to recognize that what we conceive of as
"search" applications are more complicated than a simple index that is
used to locate stuff.  We need commentary that explains that a single
application can embody multiple usage categories.