Expand on the preference overriding text
This is a light reworking of a proposal from Lila Bailey on the list.

The value of this is that it is far more comprehensive than the previous
text, which was directionally helpful, but not thorough enough to give
the right impression.  Having more text is often something to avoid, but
in this case it seems like it is adding a lot of value in terms of
making the *preferences* thing far clearer.

My only concern here -- which is a very minor one -- is that this looks
almost exclusively at the overriding of preferences to disallow.  I
couldn't come up with a convincing example of a decision to ignore
express permission.

Closes #99.