Parameters carry through
This is, of course, useless in the current state of the design.  But we
do allow for the possibility that parameters will be defined.

The only design choice here is whether to respect parameters that are
attached to invalid preferences.  The answer is that we do that, because
maybe that will turn out to be useful and there is no reason obvious
reason not to.

Closes #129.