Switch training to foundation model production
This takes the two pre-existing model training categories and changes it
into a foundation model training category.

There are a few challenges in our current definitions, particularly
around the base "AI Training" category.  As some have noted, the
definition of AI is a bit fuzzy.  This avoids that by relying on what is
now a widely-recognized definition.

That definition is not uniform, so I'm following the sage advice of a
colleague.  They suggested that making a definition for the specific use
case is best, because it allows for the definition to fit better to the
domain in which it is applied.

Some notes about aspects of "foundation model"
where this varies from accepted definitions:

* Some definitions make a point of including a measure of model size.
  US legislation says "tens of billions of parameters"; others suggest a
  firmer one billion parameter threshold.  Here, most definitions don't
  bother, except to note that the model is "large" and trained on lots of
  data.  Not being concrete seems like it will be safest.
* Many definitions refer to self-supervised learning as being typical in
  the creation of foundation models.  Because this is phrased with a
  "generally" or "typically" -- and because techniques can and do change
  -- it seemed like this wasn't a useful addition to the definition.
* This includes a note that the model has some generative capability.
  In direct contradiction to my previous note about "typically", that's
  a word I've used here.  My sense is that it is important to recognize
  this capability, particular as we are replacing the "generative AI
  training" category here.

The definition here includes fine-tuning.  This is not strictly a
foundation model thing, but my view is that a fine-tuned model is highly
likely to also be a foundation model.  Not including that in the
definition would not be consistent with the spirit of the preferences.
I acknowledge that this creates new questions and challenges -- what
about techniques like LoRA? does the name really match the scope? -- but
my view is that this is a better fit for expectations.

This does not remove any of the more general definitions that were
created to support the previous definitions (AI, training, ML, etc...).
My current view is that these are less critical now that the definition
is essentially self-contained.  We might consider removing them later.
They are certainly far less load-bearing now, so that is an easier
decision to make.  (I'm coming to the view that each category should
have a self-contained definition without relying on external
definitions, so that puts those definitions on notice.)